adapters: isolate Multiplier sgx function
Consolidate type declaration into its one file by moving the perform
method into its own file.  This produces more files, but less repeated
code and may be easier to maintain.
