+++ core/adapters/multiply_perform_sgx.go
@@ -10,38 +10,14 @@ package adapters
 import "C"
 
 import (
-	"encoding/json"
-	"fmt"
-	"strconv"
-	"unsafe"
-
 	"chainlink/core/store"
 	"chainlink/core/store/models"
-	"chainlink/core/utils"
+	"encoding/json"
+	"fmt"
+	"unsafe"
 
 	"github.com/pkg/errors"
 )
-
-// Multiplier represents the number to multiply by in Multiply adapter.
-type Multiplier float64
-
-// UnmarshalJSON implements json.Unmarshaler.
-func (m *Multiplier) UnmarshalJSON(input []byte) error {
-	input = utils.RemoveQuotes(input)
-	times, err := strconv.ParseFloat(string(input), 64)
-	if err != nil {
-		return fmt.Errorf("cannot parse into float: %s", input)
-	}
-
-	*m = Multiplier(times)
-
-	return nil
-}
-
-// Multiply holds the a number to multiply the given value by.
-type Multiply struct {
-	Times Multiplier `json:"times"`
-}
 
 // Perform returns the input's "result" field, multiplied times the adapter's
 // "times" field.
